:wrench: extract pool from next
