+++ lib/database.ts
@@ -1,13 +1,5 @@
-import mysql from "mysql2/promise";
 import { processFileName } from "./utils";
-
-export const pool = mysql.createPool({
-  host: process.env.DB_HOST,
-  port: parseInt(process.env.DB_PORT, 10),
-  user: process.env.DB_USER,
-  password: process.env.DB_PASSWORD,
-  database: process.env.DB_DATABASE,
-});
+import { pool } from "./pool";
 
 const sqlDict = {
   insertFile: "INSERT INTO mdx_files (file, file_name) VALUES (?, ?);",
