+++ lib/pool.js
@@ -1,0 +1,12 @@
+import mysql from "mysql2/promise";
+import dotenv from "dotenv";
+
+dotenv.config({ path: ".env.local" });
+
+export const pool = mysql.createPool({
+  host: process.env.DB_HOST,
+  port: parseInt(process.env.DB_PORT, 10),
+  user: process.env.DB_USER,
+  password: process.env.DB_PASSWORD,
+  database: process.env.DB_DATABASE,
+});
